+++ src/components/atoms/Image/index.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
-
-import { ImageInterface } from './interfaces';
-
 import { ImageComponent } from './styles'
+
+export interface ImageInterface {
+    src: string;
+    alt: string;
+    size?: 'small' | 'medium' | 'large';
+    isCircle?: boolean;
+    hasBorderColor?: 'PRIMARY' | 'SECONDARY';
+}
 
 function Image(props: ImageInterface) {
     return <ImageComponent {...props} />
