--- src/components/atoms/Image/interfaces.ts
@@ -1,7 +1,0 @@
-export interface ImageInterface {
-    src: string;
-    alt: string;
-    size?: 'small' | 'medium' | 'large'; 
-    isCircle?: boolean;
-    hasBorderColor?: 'PRIMARY' | 'SECONDARY';
-}
