--- src/components/atoms/icon/interfaces.ts
@@ -1,7 +1,0 @@
-import { IconSize, IconColor } from "./styles";
-
-export interface IconInterface {
-  name: string;
-  size: keyof typeof IconSize;
-  color: keyof typeof IconColor;
-}
