+++ src/components/atoms/input-text/index.tsx
@@ -1,7 +1,12 @@
-import { InputTextInterface } from "./interfaces";
-
 import { InputTextComponent } from "./styles";
+import { ColorFormat } from "@/global/theme/colors/interfaces";
 
+export interface InputTextInterface {
+    type: 'password' | 'text' | 'number' | 'email';
+    placeholder: string;
+    color?: keyof ColorFormat;
+    backgroundColor?: keyof ColorFormat | 'tranparent';
+}
 
 function InputText(props: InputTextInterface) {
     return <InputTextComponent {...props} />
