+++ src/components/molecules/Link/index.tsx
@@ -1,8 +1,6 @@
-import React,{ ReactNode }  from "react";
-
+import React, { ReactNode } from "react";
 import * as Style from './styles'
-
-import { IText } from '@/components/atoms/text'
+import { IText } from '@/components/atoms/Text'
 
 export interface ILink {
     url: string;
@@ -13,10 +11,10 @@ export interface ILink {
 
 function Link({ children, url, type, isLinkedInAnotherPage = true }: ILink) {
     return <Style.LinkComponent url={url} isLinkedInAnotherPage={isLinkedInAnotherPage}>
-            <Style.TextComponet type={type}>
-                {children}
-            </Style.TextComponet>
-        </ Style.LinkComponent>
+        <Style.TextComponet type={type}>
+            {children}
+        </Style.TextComponet>
+    </ Style.LinkComponent>
 }
 
 export default Link;
