+++ src/components/molecules/button/index.tsx
@@ -1,12 +1,21 @@
 import * as Style from './styles';
+import { ColorFormat } from '@/global/theme/colors/interfaces';
+import { IconInterface } from '../../atoms/icon/index'
 
-import { ButtonInterface } from './interfaces';
+export interface ButtonInterface {
+  size?: 'small' | 'medium' | 'large';
+  backgroundColor?: keyof ColorFormat | 'tranparent';
+  onClick: () => void;
+  isDisabled?: boolean;
+  label?: string;
+  IconProps?: IconInterface;
+}
 
-export default function Button (props: ButtonInterface) {
-    const { label, IconProps } = props;
-    return (
-      <Style.ButtonComponent {...props}>
-        {IconProps && (<Style.IconComponent  {...IconProps} color='white' size="medium"/>)}
-        {label && (<Style.TextComponent type='BUTTON' >{label}</Style.TextComponent>)}
-      </Style.ButtonComponent>)
+export default function Button(props: ButtonInterface) {
+  const { label, IconProps } = props;
+  return (
+    <Style.ButtonComponent {...props}>
+      {IconProps && (<Style.IconComponent  {...IconProps} color='white' size="medium" />)}
+      {label && (<Style.TextComponent type='BUTTON' >{label}</Style.TextComponent>)}
+    </Style.ButtonComponent>)
 }
