--- src/components/organisms/nav-bar/interfaces.ts
@@ -1,8 +1,0 @@
-import { ButtonInterface } from '../../molecules/button/interfaces'
-
-export interface NavbarInterface {
-    AboutButtonProps: ButtonInterface;
-    ProjectButtonProps: ButtonInterface;
-    ContactButtonProps: ButtonInterface;
-    ResumeButtonProps: ButtonInterface;
-}
